Move Google Analytics code from Document to App
- Document is only rendered in the server, window.location.hostname won't work
